requirements.txt
cachelib==0.10.2
click==8.1.3
dash==2.8.1
dash-core-components==2.0.0
dash-extensions==0.1.11
dash-html-components==2.0.0
dash-table==5.0.0
EditorConfig==0.12.3
Flask==2.2.2
Flask-Caching==2.0.1
imageio==2.25.0
importlib-metadata==6.0.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsbeautifier==1.14.7
kaleido==0.2.1
MarkupSafe==2.1.2
more-itertools==8.14.0
networkx==3.0
numpy==1.24.1
opencv-contrib-python==4.7.0.68
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.0
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
scikit-image==0.19.3
scipy==1.10.0
six==1.16.0
tenacity==8.1.0
tifffile==2023.2.2
Werkzeug==2.2.2
zipp==3.12.0
